requirements.txt
Django>=4.2
paho-mqtt>=1.6
